Fix refleak in type_getattr
Summary:
`_Py_bytes_isascii()` returns a `PyObject*`. Not only was this leaking a reference, but the unicode case branch was never taken.

There may be non-`PyObject*` way to do this check but a cursory glance around the code suggests there isn't. For that we'd want to duplicate/factor out the core of `_Py_bytes_isascii()` as it's fairly optimized.

Reviewed By: sinancepel

Differential Revision: D39124916

fbshipit-source-id: 018df9eb12243044de283513c47065424cafffd3