report deopt linenos using instr_offset instead of next_instr_offset for non-guard-failure deopts
Summary:
This bug got exposed due to a change in 3.10's bytecode generation where instructions that raise do not have to be followed by vacuous return instructions.

For instructions we expect to deopt, `next_instr_offset` is the instruction itself, but this is off-by-one for instructions that raise. By only using next_instr_offset when encountering a guard failure deopt, get the line numbers right, and avoid crashing due to reaching the end of the bytecode when a function ends with a `raise ...`.

Reviewed By: carljm

Differential Revision: D39063585

fbshipit-source-id: fa8caaece797cc4008c6f6d8f107b9bc41936d91