delete CO_SHADOW/NORMAL_FRAME flags
Summary:
The intent of these flags were to be able to opt in functions one-at-a-time to shadow frames before shadow frames were rolled out more widely.

Now that we've seen evidence of shadow frames working well, and plan to consolidate on them as the only mode in the future for Cinder, let's remove the per-function
ability to toggle the frame modes, and delete the related functions that operate per-function and per-module.

Reviewed By: tekknolagi

Differential Revision: D39064741

fbshipit-source-id: 1abfd22708e5281a1490d786822652750b5b06b4