adapt GetTypedArgsInfo to allow GEN_START as the first opcode of coroutines
Summary:
A new GEN_START opcode was added to Python 3.10 that must be the first opcode of coroutines. This means that CHECK_ARGS is no longer necessarily the first opcode in the bytecode of functions.

Let's relax the CHECK_ARGS check to also admit the GEN_START case for now. In the future, we might need to instead restructure bytecode generation so that CHECK_ARGS comes before GEN_START, but this diff is agnostic to that decision.

Reviewed By: carljm

Differential Revision: D39031377

fbshipit-source-id: 5d67ca9c2e3a53763f7879a422442efe618e6457