actually statically compile static dataclass generated methods
Summary:
Original Diff: D38388559 (https://github.com/facebookincubator/cinder/commit/50fcba923272204e34ba9113c63ccc16c99d849e) by sinancepel, reviewed by carljm

We custom-generate flow graphs for static dataclass generated methods. The initial implementation missed that these functions needed to be marked as statically compiled. Fix that, since these are indeed static functions we invoke against. As a corollary, we also need to fix the return type descr to appear in the flow graph as an extra const for the classloader/patching infrastructure to work, so let's add that as an API for the flow_graph construction so that it's impossible to miss :)

With this change, static dataclasses should work again in patchable cases, so re-expose static dataclasses to the world.

Reviewed By: sinancepel

Differential Revision: D39144525

fbshipit-source-id: d113840197280a0a9bdf9781565dae544a97b70f