Disable flaky test_import test
Summary:
I don't know why it was *flaky*, but using the Python compiler
(`-X usepycompiler`) adds extra frames to the traceback that are not
present in the C compiler. Since this test explicitly checks the files
in the traceback, it will fail depending on which compiler is used.
Disable the test.

Reviewed By: carljm

Differential Revision: D39144178

fbshipit-source-id: 834ebe358496843966d0fae4d35ee48abe1c5967