re-enable richards testing in Static Python
Summary:
Turns out that the Richards benchmark test wasn't working in Static Python due to an earlier refactor, coupled with a skipIf. Re-enable the test, dealing with the adjustments we needed to make due to other benchmark- related changes as we do so.

The test also revealed missing support for some shadow opcodes related to Static Python, add those.

Reviewed By: carljm

Differential Revision: D39178850

fbshipit-source-id: eb143fb30d132a5ff42de78293b408c55582c690