make base Python CodeGenerator match the stock 3.10 C compiler 1:1
Summary: The one difference our Python code generator had with the stock C compiler was that we started emitting CO_SUPPRESS_JIT in the base compiler so that we could test our Python compiler against the modified C compiler. This is breaking the baseline compiler tests, so introduce a compiler that doesn't do that, and use the base compiler when the Python we're comparing ourselves against isn't a Cinder build.

Reviewed By: jbower-fb

Differential Revision: D39214373

fbshipit-source-id: ed0efb95a097db18574b1a6bd80218914cfac9ae