enable all test_loader tests in 3.10 and fix refcount issue for Py_True/Py_False in strict modules
Summary: Because the 3.8 implementation of strict modules had access to immortal true/false, the fact that we didn't properly maintain refcount for Py_True and Py_False wasn't a problem. In 3.10, we should probably maintain the incref and let the immortalization optimization elide the modification so that the correctness isn't dependent on immortalization.

Reviewed By: carljm

Differential Revision: D39333015

fbshipit-source-id: 12d29ff70ea2bcac455f4ec7ebd2dca0997adb1b