CHECK_ARGS before GEN_START
Summary:
`3.10` includes a new `GEN_START` opcode that is placed first in
every generator function and just pops None off the stack.
Unfortunately this conflicts with Static Python, which wants a
`CHECK_ARGS` opcode first in every static function.

We need `CHECK_ARGS` first so that we can skip over it in a direct
static-to-static call, and it's actually harmless for `GEN_START` to
come after `CHECK_ARGS`, it just requires modifying one assertion
in `genobject.c`.

This diff just switches the order of the opcodes and sets us up for
actually adding the static entrypoint and doing the skip of
`CHECK_ARGS`.

Reviewed By: tekknolagi

Differential Revision: D39337115

fbshipit-source-id: d5e531b244f7fc0fed1cded0b02baed98dff00b5