Add static entry point in the interpreter
Summary:
This allows us to skip CHECK_ARGS on static invokes, which is the main thing we
care about.

The 3.8 version went a bit further and copied more of the calling code so as to be
able to eliminate checks for things we know won't be the case with static invokes
(kwargs, star-unpacked args, etc). I'm not convinced these checks are super
expensive and it's worth the added maintenance cost; we already have a task to
revisit call entry point optimizations once 3.10 is perf-testable.

Reviewed By: czardoz

Differential Revision: D38875200

fbshipit-source-id: 4b23d29a6236d1f9ffa35e54109ccf4373277c7c