Add cast to PyBytes_AS_STRING
Summary:
porting D39355623 to cinder, avoiding UBSAN errors

note that this changes in upstream 3.11 to static inline function
https://github.com/python/cpython/commit/2eca5dad0a783153318e91e3e25f16a38a632bfd

The issue is that `PyBytes_AS_STRING` expands to `(PyBytesObject *)(op))->ob_sval)` where `ob_sval` is a `char[1]`. And then call sites index the returned value with values that can be larger than 1. From what it looks, this design is not an error, likely `PyBytesObject` is used as a preamble struct in a larger memory buffer.

Fix the issue by adding an explicit cast to `(char*)`. This should be a noop.

Reviewed By: tekknolagi

Differential Revision: D39373695

fbshipit-source-id: 2eb03949ceed5557e6a7cb1684e0ff29370cb177