clarify branching and terminating HIR ops
Summary:
We had `RERAISE` listed as a branch op, which is reasonable since
it does branch, but we also make the assumption that branch ops
all have a jump target offset in their oparg, and `RERAISE` does not,
it just has a 0 or 1. This had the odd effect that we would consider
opcode at offset 1 to be a jump target (and therefore a block start)
for any function containing a `RERAISE 1`.

Remove `RERAISE` from branch opcodes and create a separate set
of block-terminating opcodes. Also consider `RETURN_PRIMITIVE`
as a block-terminating opcode (and as a return), and add a few
clarifying comments.

Reviewed By: tekknolagi

Differential Revision: D39376284

fbshipit-source-id: 4c6179b4edc586744f7267f75659ea3e3321bea2