Interpreter implementation of @native
Summary:
Backporting D39199552 to 3.8

Original summary:

As title, this diff implements the interpreter side of native
function calls, using `ctypes`. This is expected to be really slow, but
we are not focused on interpreter performance here. The goal is to make
sure that native calls _work_ in the interpreter.

Implementation:
- I've added a new opcode, `INVOKE_NATIVE`, which is emitted when
  the static compiler sees a call to a function decorated with `native`.
  - Note: earlier diffs ensure that these decorated functions cannot be
    called from non-static code. They are similar to forward declarations
    in C/C++.
- For the actual calling, there's a new classloader helper
  `_PyClassloader_InvokeNativeFunction`, and a utility function implemented
  in Python, `__static__.native_utils.invoke_native()`
- On the static compiler, there's some more functionality to ensure we
  can look up the library name. One constraint about this is, in order
  to encode the lib name in bytecode, we need it to be a `Final[str]`
  literal or a `str` constant.

In the upcoming diffs, the `invoke_native()` implementation will likely
be broken down further, since part of it (that does the symbol lookups)
needs to move into the JIT Preloader, so we can burn in native symbol
addresses at compile time.

Reviewed By: tekknolagi

Differential Revision: D39391224

fbshipit-source-id: 421906d0f591f9b3e7b780e9899daa16c46acddc