Import cpython_lldb into cinder
Summary:
From: https://github.com/malor/cpython-lldb
Which is MIT licensed: https://github.com/malor/cpython-lldb/blob/master/LICENSE

This distributes it the same way gdb extensions are distributed.

The idea with this is to distribute it in downstream binaries the same way as gdb.py, so `$binary-lldb.py`.

LLDB supports importing modules, but I think it may be easier for a debugger user to find `$binary-lldb.py` than a path deeply embedded in site modules.

Thoughts? Open to suggestions.

After reading into this code, it appears it already attempts to find python frames in cpu registers, which is what the additions I added into our gdb ext did.

Couple of differences from `py-walk-*`:
- `py-walk-stacks` finds the first frame and walks it upwards rather than try to find every frame in the native stack, which is a bit faster and less fragile.
- I miss some the more terse output of `py-walk-threads` when dumping python stacks from all threads.

But these are all easy enough to change later.

Reviewed By: carljm

Differential Revision: D39418964

fbshipit-source-id: f39a12726fb192cebb7eaf2992cbbb80c494bb62