test with lazy imports and without pycs
Summary:
Fix the py-compiler in case it is used with lazy imports enabled
and without any pyc files in place for the py-compiler. In this case
we need to ensure that the necessary parts of the py-compiler are
all eagerly imported (and thus compiled with the C compiler) when it
is installed, otherwise we will later hit failures when it tries to compile
parts of itself, which obviously doesn't work.

The simplest way to ensure the compiler is fully loaded when
installed is to temporarily turn off global lazy imports while importing
it. Todo this for the strict/static compiler requires moving its
installation from `site.py` into `initconfig.c`; it's nice to have all three
loader installations handled in the same place anyway.

Maybe in the future they should all be moved out of C and into
`site.py`, but this will require better Python API around lazy imports,
which comes in later Python versions.

Reviewed By: czardoz

Differential Revision: D39455163

fbshipit-source-id: 0d110084fb558e0815e898c3c62f17d45c7a7fc2