Revert "Fix race condition in Runtime::allocateCodeRuntime"
Summary:
This reverts D39177666. Based on the ASAN error it was meant to fix,
the real issue is that we were entering the JIT during `_PyJIT_Finalize()`, not
anything to do with threading (`SlabAllocator::allocate()` is already
thread-safe anyway). I'll put up a fix for the underlying issue after this
diff.

Reviewed By: tekknolagi

Differential Revision: D39518847

fbshipit-source-id: 311c673c3a69ed8d06020a483306afb2239a460b