Port eager execution of coroutines
Summary:
This task turns on eager execution of coroutines. If a called function reaches a `return` without ever needing to `await`, we return that value directly without ever creating a coroutine object or deferring to the event loop. (A coroutine object exists largely to store frame state. If we have no frame state to store, we don't need to create a coroutine.) In contrast, vanilla CPython will naively create a coroutine object in every case.

At this point, we eagerly run codepaths that:
* don't involve IO since IO always suspends execution
* don't use `asyncio.gather` (`gather` always suspends execution even if all constituents can be finished eagerly)

Further diffs will expand the number of codepaths for which eager execution is valid.

The major changes introduced by this diff are:
* Check if the next instruction is `GET_AWAITABLE` to decide if call is awaited
* Use a `nargs` parameter as a facility to carry the bit to indicate if call is awaited. Vectorcalls use similar way to pass around extra information (but using different bit) so they can be combined together without big issues
* Return the result of eagerly evaluated coroutine as tagged pointer with a LSB being set (this only happen if caller indicates that call is awaited). These results can be either:
    * coroutine object with coro_waiter field being set to denote that coroutine was suspended and coro_waiter is a future that should be fulfilled to proceed the coroutine
    * anything else - returned as result of eager execution

This diff is based off of the 3.7 -> 3.8 port D23310054.

Reviewed By: sinancepel

Differential Revision: D39102216

fbshipit-source-id: 5ee13bea6c994757efb904c707c1a43a2dc4760e