test with lazy imports and without pycs
Summary:
Port of D39455163 (https://github.com/facebookincubator/cinder/commit/1c78f3147ab0c65b770c5245796961b078308c22). This required some adjustment for 3.10, so the diff should
be reviewed as a new diff. Original summary (lightly edited):

Fix the py-compiler in case it is used with lazy imports enabled
and without any pyc files in place for the py-compiler. In this case
we need to ensure that the necessary parts of the py-compiler are
all eagerly imported (and thus compiled with the C compiler) when it
is installed, otherwise we will later hit failures when it tries to compile
parts of itself, which obviously doesn't work.

The simplest way to ensure the compiler is fully loaded when installed
is to temporarily turn off global lazy imports while importing it. To do
this for the strict/static compiler requires moving its installation
from `site.py` into `initconfig.c`; it's nice to have both loader
installations handled in the same place anyway.

Maybe in the future they should all be moved out of C and into
`site.py`, but this will require better Python API around lazy imports,
which needs backporting from the PR against Python 3.12.

Reviewed By: Kronuz

Differential Revision: D39523145

fbshipit-source-id: 0a8480016ca7038e7c359769afd8e7f980c9bbec