do not attempt to mark non-static callables in static types as static
Summary:
We had a bug which manifested itself as a spurious "missing type annotation on static compiled method" error - the root cause is that static types can have non-static methods due to decorators which we do not understand.

When constructing vtables, we previously assumed that if we got a function from a static type's dict, we were good to use that as an original static base. This is not necessarily true, so insert a check in the classloader for getting an original.

Reviewed By: tekknolagi

Differential Revision: D39475693

fbshipit-source-id: 3675d4cb9f2932d3af905987ee6318085740dc69