port "[static] do not attempt to mark non-static callables in static types as static"
Summary:
Original diff: D39475693 (https://github.com/facebookincubator/cinder/commit/df7887644d80038cf9240ed5fa836f3d74268f9c)

Original summary:

We had a bug which manifested itself as a spurious "missing type annotation on static compiled method" error - the root cause is that static types can have non-static methods due to decorators which we do not understand.

When constructing vtables, we previously assumed that if we got a function from a static type's dict, we were good to use that as an original static base. This is not necessarily true, so insert a check in the classloader for getting an original.

Reviewed By: tekknolagi

Differential Revision: D39555824

fbshipit-source-id: 025f8119c99dd0f6d5ec55a54ff07b47e3a5bbae