Default PyCodeObject.co_qualname to co_name
Summary:
We do this in [3.8] and I think not including it the 3.10 port so far
was just an oversight. Change all existing writes to `co->co_qualname` in
`codeobject.c` to use `Py_XSETREF`, to avoid leaking any previous value.

I also bumped the bytecode magic number and ran `make regen-importlib`

[3.8]: https://github.com/facebookincubator/cinder/blob/cinder/3.8/Python/compile.c#L6213-L6217

Reviewed By: itamaro

Differential Revision: D39586122

fbshipit-source-id: 7d7d3ad58f5a7f4412d9d59228169248c5732181