Fix potential use-after-free during preloading
Summary:
We store borrowed references to units that are to be preloaded (https://fburl.com/code/1aj7g8m1)
or compiled (https://fburl.com/code/u4mqcu4o), however, preloading can execute arbitary Python code
which may delete yet-to-be-processed units. We had nothing in place to handle this scenario and were
hitting use-after-frees in production when we tried to preload deleted units. To remedy  this we keep
track of units that are deleted during preloading and skip preloading/compiling them.

Reviewed By: carljm

Differential Revision: D39806460

fbshipit-source-id: 25ec649b3d4b13cacdb5afb95606044ca18fc53b