Fix Ref<> bug in get_jit_list()
Summary:
`g_jit_list->getList()` already returns a `Ref<>`. Passing a `Ref<>`
to `Ref<>::steal()` triggers the implicit conversion to a borrowed `PyObject*`,
which we then steal a reference from, causing over-decrefing when all the
relevant `Ref<>`s are destroyed.

Long-term, we should get rid of this implicit borrowing since there are too
many ways for it to go wrong. For now, we can prevent this specific mistake by
deleting the right overload of `Ref<>::steal()`.

Reviewed By: alexmalyshev

Differential Revision: D39836227

fbshipit-source-id: 88cd2d8c748ff712383038598763b58fa64224fc