Port `Fix potential use-after-free during preloading`
Summary:
This is a mostly clean cherry pick of D39806460 (https://github.com/facebookincubator/cinder/commit/8992c12c5fda538ddff3cd7eebbb0f2e918effb0). The only conflicts
were in `test_cinderjit.py` and were largely due to the better support
for checking line numbers that's in 3.10 but not in 3.8.

Reviewed By: swtaarrs

Differential Revision: D39854530

fbshipit-source-id: 331a5023b53f439f340418ad7c152c301fb70edb