add env var to disable comprehension inlining
Summary:
The Numba JIT does not allow compiling functions with
`DELETE_FAST` opcode in their bytecode. Our comprehension
inlining deletes the comprehension loop variable after running the
comprehension (for name-visibility compat with normal
comprehension-as-function behavior.) So any function compiled
by Cinder, with a comprehension in it, will fail Numba JITting.

Comprehension inlining is a significant perf win in IG prod, so we
don't want to simply remove it.

This diff is a quick hack to add an environment variable,
`PYTHONNOINLINECOMPREHENSIONS`, which will disable
comprehension inlining in the compiler. This will unblock Cinder
use with Numba JIT in the short term while we figure out a better
plan.

This is a hack because the compiler should not be environment-
sensitive, and in particular any outside input that changes
compiler output ideally should be reflected in the pyc file name.
Because we are not doing that in this diff (it would bring a whole
raft of other complications and problems), if someone runs some
Python code with this env var set, and then runs the same code
without this env var set (without changing the code), the previous
pyc files will be used and the effect of changing the env var will
not be seen.

But in practice the way this will be used in our builds, that case
(switching the env var on and off while reusing the same pyc
files) should not occur, so we can live with this hack for now.

I went with an environment variable rather than a `-X` flag
because in any scenario where this is used, it is likely we want it
in any subprocesses also.

Reviewed By: mpage

Differential Revision: D40001105

fbshipit-source-id: cab3d33deb5e6ea4dc8b6688e0eb25f1d0f1ec62