Fixing tests for Lazy Imports
Summary:
This fixes tests for Lazy Imports. The changes here are backported from main branch, which fixes several imports and issues that become apparent with Lazy Imports enabled. Some tests were not importing things properly or using things in a way that is Lazy Imports compatible. This fixes some of the issues.

**Note:** This is a backport of the tests fixed for Lazy Imports in the main branch

Reviewed By: itamaro

Differential Revision: D40034673

fbshipit-source-id: 8a3d89da6bcb8f1f8f62dc3dc5186ee435712e0b