Bring over missing heap immortalization tests
Summary:
Heap immortalization was already ported but looks like these tests were missing. Also move a Python callable function for test only to `gcmodule.c` from `cinder.c` - putting it with the rest of the code. This seems to be needed now as 3.10 seems more inclined to disallow symbolic dependencies between modules.

I've also moved the test from `test_gc` where it's kind of hard to see it's actually part of Cinder to `test_cinder`. Really this whole thing should either be upstreamed or live in a Cinder module.

Reviewed By: sinancepel

Differential Revision: D40051491

fbshipit-source-id: b53b92aefbb3f60d7a041f18167fc8638ddbf733