Fix eager coroutine dispatch for Static Python -> Static Python
Summary:
The awaited flag was being zeroed out in this case to facilitate porting. Also turned out the existing tests didn't exercise the Static Python -> Static Python path so I updated it. Unfortunately, this is a little hard to *prove* in the test itself but I verified it using a debugger and checking to see `_PyFunction_CallStatic` gets invoked.

Also stopped compiling Static Python code twice in `EagerCoroutineDispatch`. Not sure why I did that in the first place.

Reviewed By: itamaro

Differential Revision: D40127948

fbshipit-source-id: bf910231160e07e296f34477eef754c89658e349