Annotate some builtin C functions with return types
Summary:
The JIT knows the exact return types of some functions. For
example, `str.lower` will (if it succeeds) always return a `str`. Even
it's called on a `str` subclass. Load the `PyCFunction`s off the static
`builtins` module and other builtin types to guard against redefining
`builtins`.

Also take the builtins initialization out of `DynamicComparisonElimination` and
lazily initialize it once in `Runtime`.

Reviewed By: jbower-fb

Differential Revision: D39984428

fbshipit-source-id: bed3fd1971d6b717fd9e73e17c6a2e9edb49a0e4