Make PrimitiveBox<TCBool> not deopt
Summary:
Unlike the other `PrimitiveBox` variants, this one can never
deopt -- it does not allocate and just returns one of `Py_True` and
`Py_False`. Also mark this as a borrowed reference and let
`RefcountInsertion` deal with the refcounting for us. This makes
followup diffs (like using `cmov` for this implementation) simpler.

Reviewed By: sinancepel

Differential Revision: D37401142

fbshipit-source-id: 139dc21dbb4eeff86d8505e0fece3793d0b2728c