Optimize PrimitiveBox TCBool + fix Select bugs
Summary:
Boxing a TCBool is easy because it's just a matter of returning Py_True or
Py_False.  This should be more efficient than calling out to a runtime helper
(JITRT_BoxBool).

However this exposed issues with the Select LIR instruction. It currently takes
one dest register, one source register, and two immediates.  It clobbers the
source register, which works today because the one callsite uses a temporary for
the source, but when I tried to use it with a real value things went awry.  I'm
updating Select to take its third argument by register so we don't run the risk
of clobbering anymore.

The other issue that I hit is that Select is marked as always using 64-bit
inputs, which doesn't work when we try to select based on a byte value (TCBool).
I'm setting it to kDefault.

Reviewed By: tekknolagi

Differential Revision: D40145773

fbshipit-source-id: 9ad254461eaf3bcd4a94feb9b36e402ac8a329f7