Use X Macro to define BinaryOpKind
Summary:
There's duplication between the list of enum values and the array of their
string names, and we've already introduced bugs as a result of it.  See how
kNumBinaryOpKinds is one less than kPowerUnsigned, and how the array of string
names doesn't include "PowerUnsigned".

Putting this type of change up to get peoples' opinions.  I'll do the same for
the other subopcodes if folks like it.

Reviewed By: tekknolagi

Differential Revision: D40266485

fbshipit-source-id: b8bb255fd803ced5cfccf1b5a3d895accb50495a