Use std::string_view for jit::hir::Type::parse()
Summary:
This doesn't need to take a full `std::string`, it can work with
`std::string_view` instead which is cheaper to work with.

I needed to add a helper for parsing string views into integers and along the
way I found more places in the HIR parser where to apply it.

Reviewed By: tekknolagi

Differential Revision: D40307934

fbshipit-source-id: 088c544cccd4fc3aa777cac2a73ddc6b596cc02d