Move PrimitiveCompare optimization out of simplifyIntConvert()
Summary:
There was code in `simplifyIntConvert()` that was really optimizing a
`PrimitiveCompare`, but only when its output happened to be an input to
`IntConvert`. Moving that logic to `simplifyPrimitiveCompare()` will let it
apply in all situations.

I also removed the `IntConvert` and `PrimitiveBox` from the relevant `Simplify`
test since they're no longer necessary to trigger the optimization, and
clarified the name of the test.

Reviewed By: jbower-fb

Differential Revision: D40313955

fbshipit-source-id: df06abda03c40b33f65461ff2ea67258314f422c