Add hir::PrimitiveBoxBool
Summary:
`PrimitiveBox<CBool>` had enough special cases (deopt, memory effects,
return type) that it should really be a separate HIR opcode. This means that
people who want to emit a `PrimitiveBox` of a value that could be a `CBool`
will have to check and use a different opcode, but I think it leaves us with
fewer surprises and less complexity overall. Importantly, we again have the
property that `instr->asDeoptBase() == dynamic_cast<DeoptBase*>(instr)`,
avoiding one route for subtle bugs (we can separately discuss whether or not to
use inheritance for deopt like this).

Reviewed By: tekknolagi

Differential Revision: D40313972

fbshipit-source-id: 80ec6c32e6517fdade48775af460eb666908d563