Mark value types in type descriptors
Summary:
Ports D39578828 (https://github.com/facebookincubator/cinder/commit/3fca283a000b821aa4cd6c0f0dd69474c0021d35). Omits the static entry point changes since we haven't brought the complex logic over yet (see D38875200's summary for context as to why).

Original summary:

This marks value types in our descriptors with a '#' marker.  This allows us to avoid doing any type resolution on them to prevent additional module loads from occurring in surprising places.

Without this we can trigger a type load in a couple of different spots unexpectedly.  One is after a function definition of a function which is not on the JIT list.  We'll eagerly initialize the functions entry point (to avoid COW on the function later) and need to resolve the return type.  This triggers a type load, which in turn could trigger a module load.

The other place is in initializing fields.  The original design was that the types of fields would be lazily resolved except for primitives where we need to know the exact type to calculate the size.  But over time (possibly with the introduction of type codes) we've switched to just resolving the whole type.  This again triggers a type load, which in turn can trigger a module load.

The initial reason for wanting to have a lazy design here is that we allow more cross-module references, references to types defined later in the module, etc...
With the introduction of lazy imports we can sometimes still successfully load despite circular references, it can just take a long time.  In one worst case scenario we were seeing a module get loaded 50 times and subsequently triggering the JIT of a large method within it 50 times as well, leading to a large amount of memory usage from the JITed methods hanging around.

Reviewed By: carljm

Differential Revision: D40354539

fbshipit-source-id: b282e944c3e4933ba6b61847948719cf339e96bd