Port asyncio.gather optimizations + AwaitableValue + ContextAwareTask features
Summary:
Primarily ports a native implementation of `asyncio.gather()` which utilizes the `Ci_Py_AWAITED_CALL_MARKER` to eagerly execute coroutines. It also fast-paths `AwaitableValue`, `AsyncLazyValue`, and completed futures.

Also in this diff is the implementation of `AwaitableValue`, and some new features for `ContextAwareTask`. All these are somewhat intertwined with the native gather implementation so we're just lumping them all into this diff.

Reviewed By: itamaro

Differential Revision: D39545755

fbshipit-source-id: a99a92a77b842504a217d559fcbc18b39c0c0ec7