Add _PyDict_HasUnsafeKeys()
Summary:
Port of D35732041.

Original Summary:
This returns true if the given dict has any deferred values, or any
keys that are not exactly `str` (it also returns true if anyone has even looked
up in the dict with a non-`str` key; I'm hoping that isn't a problem for our
use case).

The target use case is for looking up items in the JIT, where we need to avoid
Python heap mutations (including arbitrary code execution).

Reviewed By: tekknolagi

Differential Revision: D40196536

fbshipit-source-id: 886ba715c70b33a492b771e39ffa236a64302b25