Port: Disable JIT earlier in JIT shutdown
Summary:
Port of D39518824.

Original Summary:
We were seeing an ASAN use-after-free error caused by
`Runtime::releaseReferences()` invoking an object destructor, which tried to
reenter Python, which tried to JIT-compile a function, which allocated a
`CodeRuntime`, all while `Runtime::releaseReferences()` was iterating over its
list of `CodeRuntime`s. This all amounts to mutating a container while
iterating over it, causing undefined behavior (use-after-free, in this case).

I think the best fix for this is to disable the JIT at the beginning of
`_PyJIT_Finalize()`, so we can tear down the various JIT data structures with
no risk of something trying to reenter the JIT. It's still possible for
arbitrary user code to run in an object destructor, but that should stay in the
interpreter and remain harmless.

Reviewed By: alexmalyshev

Differential Revision: D40769826

fbshipit-source-id: abd8fc4800f083c507fa174896c24f058fd4080f