add __spec__ and __loader__ attributes to _static module
Summary:
All modules should have `__spec__` and `__loader__` attributes, and there's an
importlib test that tests all of `sys.modules` for this.

Our `_static` built-in support module for Static Python did not have these attributes.
This caused a flaky test failure in importlib iff the SP tests were run before the
importlib tests in the same process. This diff fixes that.

Reviewed By: tekknolagi

Differential Revision: D40920606

fbshipit-source-id: bf19144ec26130890a76c8ed47390cb0fde0f910