Allow overloading recognized decorators
Summary:
Currently, the overload decorator doesn't allow overloads of
classmethods and staticmethods. This change gets them to be allowed.

Note: There's a different but related issue, where we don't allow
overloads for unrecognized decorators. That one's a little gnarly to
fix, because our policy for unknown decorators is that we don't process
any of them at all.

If we were to fix that, it'd perhaps involve trying to recognize the overload
decorator, in `Function().finish_bind()`, and setting a flag on
`UnknownDecoratedMethod`s. That's not a pressing matter though.

Reviewed By: carljm

Differential Revision: D40881327

fbshipit-source-id: a2f552351ed62f3f2a0d6a2d70d14ef2bf1b5eca