add __spec__ and __loader__ attributes to _static module
Summary:
Port of D40920606. Original summary:

All modules should have `__spec__` and `__loader__` attributes, and there's an
importlib test that tests all of `sys.modules` for this.

Our `_static` built-in support module for Static Python did not have these attributes.
This caused a flaky test failure in importlib iff the SP tests were run before the
importlib tests in the same process. This diff fixes that.

Reviewed By: tekknolagi

Differential Revision: D40981442

fbshipit-source-id: 87958595df1df641df65953f85f86f0497f1a97f