Use the AST optimizer to get rid of dead code
Summary:
This small change helps get rid of dead code gated by
sys.hexversion checks.

We do not want to keep this around for long. But Static Python doesn't
do well when having multiple class definitions for the same class,
because it cannot determine which one will be used at runtime, so it
might as well just be dynamic.

In this case, we can safely recognize and remove some paths as dead
code.

Reviewed By: carljm

Differential Revision: D40911543

fbshipit-source-id: f11f03040ee41cf698c4422633e895081e234a10