don't forget globals caches
Summary:
When we disable a globals cache (e.g. because the globals or builtins dict
became unwatchable), we currently remove the `GlobalCacheValue` from the
`global_caches_` map. This can result in a `GlobalCache` instance having a
`pair_` pointer that is now a dangling pointer to freed memory.

We could potentially work with this, by trying to ensure that we never hold onto
a `GlobalCache` instance around anything that could execute arbitrary Python
code. But given that under lazy imports `PyDict_GetItem` can execute arbitrary
Python, this would require some significant refactoring of the globals caching
implementation.

An alternative approach, given that globals/builtins becoming unwatchable
shouldn't be all that common anyway, is to just stop freeing caches. Then `pair_`
will never be a dangling pointer. This isn't an issue for correctness, since we
still invalidate the cache by setting the cached pointer to null; we just forgo a
(very small) memory savings in a rare edge case.

Reviewed By: tekknolagi

Differential Revision: D40956578

fbshipit-source-id: 4866f212a078bf25a338d0f8ef90bbdae1275013