backport 'don't raise on excessive stack consumption'
Summary:
3.10 added a new check in the compiler to error if a maximum
stack size was exceeded, on the (mistaken) assumption that all
cases in the compiler that could lead to very large stack sizes
had been fixed. Later it was realized that large stack sizes were
still possible, so this check was reverted (in 3.10.6+, 3.11, and
3.12.) Backport the revert to our 3.10 (which is based on 3.10.5.)

There might be an argument for updating our 3.10 to the latest
upstream 3.10 branch (or latest 3.10 minor release) to bring in
more of these bugfixes and compatibility fixes before we have to
hit them ourselves. But this might also bring conflict-resolution
hassles for bugs we wouldn't even be affected by. For now, just
backport this one fix.

Also submitted https://github.com/python/cpython/pull/99237
upstream to ensure this case isn't missed if/when the max stack
size check is re-added in 3.12.

Reviewed By: itamaro

Differential Revision: D41111919

fbshipit-source-id: 5480357678fef562942693ab9521b07c64bb86b4