Fix strict decorators in no-analyze mode
Summary:
Currently, when strict analysis is disabled (either using PYTHONSTRICTDISABLEANALYSIS, or using `from __strict__ import allow_side_effects`), we skip the analysis step.

However there's a dependency between strict analysis and the AST rewriter pass (and eventually our codegen). During analysis, we build a mapping of AST nodes and add extra decorators to the AST. So, the following code:

```
nutritious
class C: pass
```

With strict analysis, has the AST:
```
    ClassDef(
      ...
      decorator_list=[
        Name(id='mutable', ctx=Load()),
        Name(id='<mutable>', ctx=Load())])],
```

Without analysis:
```
    ClassDef(
      ...
      decorator_list=[
        Name(id='mutable', ctx=Load())])],
```

The `<mutable>` decorator, added by the Strict Modules AST pre-processor, is currently used in strict-modules codegen as an identifier to indicate whether a class should or should not be immutable.

In this change, we do the simplest thing possible, which is relax the checks to also look at the un-processed names of the decorators.

The downside of this change is: it could create false positives in the case where we have naming clashes.

A potential alternative solution is to change the `mutable` decorator to unset the `Ci_Py_TPFLAGS_FROZEN` flag. The complication there is, type freezing happens just before module execution finishes (so after class decorators run). We'd have to figure out a way to change that order.

Reviewed By: carljm

Differential Revision: D41051080

fbshipit-source-id: de9a8e064e162a3f5cbe3ee143da6da5082bd2a2