Clean up inputs_live_across LIR instruction flag
Summary:
This flag should only be set when codegen for the instruction needs to
write to the output register before reading all of the input registers. This
doesn't apply to yields or deopting instructions: they use `DeoptMetadata`
to track which values are live in which physical locations, but that's orthogonal
to whether or not there should be overlap between the liveness of the inputs
and output.

I audited all existing codegen rules in `autogen.cpp` and left it set for the
instructions that really need it: `Sub`, `Fsub`, `Fdiv`, and `Select`. As part
of this, I ran into and fixed a latent bug with `Fsub`, where the rewrite pass
used to emit `Negate` of a double, which we don't support.

I also added some comments to the codegen for other binary ops explaining why
it's OK for their codegen to write the output before reading all inputs even
though they don't have `inputs_live_across` set.

Reviewed By: mpage

Differential Revision: D41005913

fbshipit-source-id: 0a5a20e6ef1956236777fd88c2b1230251ed15d7